Completed implementation of repository pattern for regions and added AutoMapper

+++ NZWalksAPI/Repositories/SQLRegionRepository.cs
@@ -11,23 +11,29 @@ namespace NZWalksAPI.Repositories
         {
             _dbContext = context;
         }
-        public async Task<Region> Create(Region region)
+        public async Task<Region> CreateAsync(Region region)
         {
             await _dbContext.Regions.AddAsync(region);
             await _dbContext.SaveChangesAsync();
             return region;
         }
 
-        public async Task Delete(Region region)
+        public async Task<Boolean> RemoveAsync(Guid id)
         {
+            Region? region = await _dbContext.Regions.FirstOrDefaultAsync(x => x.Id == id);
+            if (region == null)
+            {
+                return false;
+            }
             _dbContext.Regions.Remove(region);
             await _dbContext.SaveChangesAsync();
+
+            return true;
         }
 
-        public async Task<Region> Get(Guid regionId)
+        public async Task<Region?> GetByIdAsync(Guid regionId)
         {
-            Region region = await _dbContext.Regions.FirstOrDefaultAsync(u => u.Id == regionId);
-            return region;
+            return await _dbContext.Regions.FirstOrDefaultAsync(x => x.Id == regionId);
         }
 
         public async Task<List<Region>> GetAllAsync()
@@ -36,10 +42,18 @@ namespace NZWalksAPI.Repositories
             return regions;
         }
 
-        public async Task<Region> Update(Region region)
+        public async Task<Region?> UpdateAsync(Guid id, Region region)
         {
+            Region? existingRegion = await _dbContext.Regions.FirstOrDefaultAsync(x => x.Id == id);
+            if (existingRegion == null)
+            {
+                return null;
+            }
+            existingRegion.Name = region.Name;
+            existingRegion.Code = region.Code;
+            existingRegion.RegionImageUrl = region.RegionImageUrl;
             await _dbContext.SaveChangesAsync();
-            return region;
+            return existingRegion;
         }
     }
 }
